requirements.txt
asgiref==3.4.1
cached-property==1.5.2
Django==3.2.8
django-localflavor==3.1
django-url-filter==0.3.15
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
enum-compat==0.0.3
PyJWT==2.3.0
python-stdnum==1.17
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
